Make TestMessageSerialization robust to missed messages (#456)
Though unlikely, I think it's possible for the first couple messages to be missed since we directly
start publishing after creating the subscription. If this happens, it will throw off the expected
value based on the counter. I think this is what was responsible for the reported test failure in #451.

This change accounts for possible missed messages by setting the counter to the value in the first received message.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>