Simplify the test_secure_subscriber code. (#471)
The main reason to do this is that we were sometimes running out
of memory on the buildfarm while building this code.  I suspect
that it is because of the templated attempt_subscriber() functions
that had to get expanded for quite a few different messages.
However, it turns out that we were only ever using 2 of those
messages in tests themselves.  Simplify this whole thing by
only creating templated functions for the messages we are
actually going to use.

Similarly, simplify the quite complicated CMake logic a bit to
only have a list of the messages that we are going to actually
use.  If we want to add more messages in the future, we can
always expand this list again.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit 8fc14b48bbf3f102fa6b19443c90d1841aca3b9a)