Use test fixtures to create SROS artifacts
The current implementation creates SROS artifacts during the configure
phase. This is undesirable for at least a variety of reasons, such as:
* The ros2 binary is invoked during build, meaning that a regression in
  that tool will break the entire CI build instead of appearing as a
  test failure.
* Some of the environment is captured during the artifact generation
  and it may not be the same when the tests are invoked. Namely,
  changes to ROS_DOMAIN_ID between build and test will break.

The test fixtures work similarly to tests themselves, and are
automatically run if any tests requiring them are enabled. If the
fixture fails, the dependent tests are not attempted.

Signed-off-by: Scott K Logan <logans@cottsay.net>